requirements.txt
pycryptodome
numpy
pandas
scikit-learn
matplotlib
